local patches

+++ casadi/core/matrix/sparsity.cpp
@@ -123,7 +123,7 @@ namespace casadi {
     return (*this)->size2();
   }
 
-  int Sparsity::numel() const {
+  size_t Sparsity::numel() const {
     return (*this)->numel();
   }
